requirements.txt
fastapi
uvicorn
anthropic
supabase
python-dotenv
feedparser
httpx
trafilatura
sentence-transformers
numpy
